Fix miss-parsing for poorly formatted text
When using the spacy integration, an improperly transcribed piece of text cased the parsing to fail. In this case it was "y. Million" (should have been "why millions").  This had a pattern match, but group_1 was `. `which can't be parsed to a number.

This PR just allows us to gracefully fall back to the best effort in that case.